// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
const defaultOptions = {};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type AttributesEntity = {
  __typename?: "AttributesEntity";
  traitType?: Maybe<Scalars["String"]>;
  value?: Maybe<Scalars["String"]>;
};

export type AttributesEntityInput = {
  traitType?: Maybe<Scalars["String"]>;
  value?: Maybe<Scalars["String"]>;
};

export type AuthPayload = {
  __typename?: "AuthPayload";
  token: Scalars["String"];
  user: User;
};

export type Collection = {
  __typename?: "Collection";
  family?: Maybe<Scalars["String"]>;
  name?: Maybe<Scalars["String"]>;
};

export type CollectionInput = {
  family?: Maybe<Scalars["String"]>;
  name?: Maybe<Scalars["String"]>;
};

export type Comment = {
  __typename?: "Comment";
  createdAt?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  isCommentMine: Scalars["Boolean"];
  text: Scalars["String"];
  tweet?: Maybe<Tweet>;
  updatedAt?: Maybe<Scalars["String"]>;
  user?: Maybe<User>;
};

export type CreatorsEntity = {
  __typename?: "CreatorsEntity";
  address?: Maybe<Scalars["String"]>;
  share?: Maybe<Scalars["Int"]>;
};

export type CreatorsEntityInput = {
  address?: Maybe<Scalars["String"]>;
  share?: Maybe<Scalars["Int"]>;
};

export type File = {
  __typename?: "File";
  id: Scalars["ID"];
  tweet?: Maybe<Tweet>;
  url: Scalars["String"];
  user?: Maybe<User>;
};

export type FilesEntity = {
  __typename?: "FilesEntity";
  type?: Maybe<Scalars["String"]>;
  uri?: Maybe<Scalars["String"]>;
};

export type FilesEntityInput = {
  type?: Maybe<Scalars["String"]>;
  uri?: Maybe<Scalars["String"]>;
};

export type Gif = {
  __typename?: "Gif";
  fixedHeightUrl: Scalars["String"];
  id: Scalars["ID"];
  originalUrl: Scalars["String"];
  title: Scalars["String"];
  tweet?: Maybe<Tweet>;
};
// ...
export type Nft = {
  __typename?: "NFT";
  attributes?: Maybe<Array<Maybe<AttributesEntity>>>;
  collection?: Maybe<Collection>;
  description?: Maybe<Scalars["String"]>;
  externalUrl?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  image: Scalars["String"];
  name?: Maybe<Scalars["String"]>;
  properties?: Maybe<Properties>;
  publicKey: Scalars["String"];
  sellerFeeBasisPoints?: Maybe<Scalars["Int"]>;
  symbol?: Maybe<Scalars["String"]>;
};
// ...
export type Properties = {
  __typename?: "Properties";
  category?: Maybe<Scalars["String"]>;
  creators?: Maybe<Array<Maybe<CreatorsEntity>>>;
  files?: Maybe<Array<Maybe<FilesEntity>>>;
};
// ...
export type Reaction = {
  __typename?: "Reaction";
  count: Scalars["Int"];
  emojiId: Scalars["String"];
  id: Scalars["ID"];
  isMine: Scalars["Boolean"];
  skin?: Maybe<Scalars["Int"]>;
  tweet: Tweet;
  user: User;
};

export type Retweet = {
  __typename?: "Retweet";
  createdAt: Scalars["String"];
  id: Scalars["ID"];
  tweet?: Maybe<Tweet>;
  updatedAt: Scalars["String"];
  user?: Maybe<User>;
};

export type Tip = {
  __typename?: "Tip";
  createdAt?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  tipAmount: Scalars["String"];
  tweetId: Scalars["String"];
  updatedAt?: Maybe<Scalars["String"]>;
  userId: Scalars["String"];
};

export type TipInput = {
  tipAmount: Scalars["String"];
  tweetId: Scalars["String"];
  userId: Scalars["String"];
};

export type Tweet = {
  __typename?: "Tweet";
  childTweets?: Maybe<Array<Maybe<Tweet>>>;
  comments: Array<Comment>;
  commentsCount: Scalars["Int"];
  createdAt?: Maybe<Scalars["String"]>;
  files: Array<File>;
  gif?: Maybe<Gif>;
  id: Scalars["ID"];
  isRetweet: Scalars["Boolean"];
  isTweetMine: Scalars["Boolean"];
  mentions?: Maybe<Array<Scalars["String"]>>;
  nft?: Maybe<Nft>;
  parentTweet?: Maybe<Tweet>;
  reactions: Array<Reaction>;
  retweets: Array<Retweet>;
  retweetsCount: Scalars["Int"];
  tags: Array<Scalars["String"]>;
  text: Scalars["String"];
  tipsCount?: Maybe<Scalars["String"]>;
  updatedAt?: Maybe<Scalars["String"]>;
  user?: Maybe<User>;
};

export type User = {
  __typename?: "User";
  avatar: Scalars["String"];
  bio?: Maybe<Scalars["String"]>;
  comments: Array<Comment>;
  consumerName?: Maybe<Scalars["String"]>;
  coverPhoto?: Maybe<Scalars["String"]>;
  createdAt?: Maybe<Scalars["String"]>;
  dob?: Maybe<Scalars["String"]>;
  email?: Maybe<Scalars["String"]>;
  files: Array<File>;
  followers: Array<User>;
  followersCount: Scalars["Int"];
  following: Array<User>;
  followingCount: Scalars["Int"];
  handle: Scalars["String"];
  id: Scalars["ID"];
  isFollowing: Scalars["Boolean"];
  isSelf: Scalars["Boolean"];
  location?: Maybe<Scalars["String"]>;
  newMentionsCount: Scalars["Int"];
  nonce: Scalars["String"];
  publicAddress: Scalars["String"];
  reactions: Array<Reaction>;
  retweets: Array<Retweet>;
  tweets: Array<Tweet>;
  tweetsCount: Scalars["Int"];
  updatedAt?: Maybe<Scalars["String"]>;
  website?: Maybe<Scalars["String"]>;
};